requirements.txt
python-binance
requests
typer
rich
python-dotenv
